Removed unnecessary functionality from rvsteel-api.c and renamed it to rvsteel-soc.c
